Removing bulma.py, and moving that input into layout with the others.
