+++ crispy_bulma/layout.py
@@ -3,10 +3,12 @@ from crispy_forms.layout import (
     Fieldset, HTML, Hidden, Layout, MultiField,
     MultiWidgetField
 )
+from crispy_forms.utils import TEMPLATE_PACK, render_field
 
 __all__ = [
     # Defined in this file
     "Button", "Column", "Reset", "Row", "Submit",
+    "IconField",
 
     # Imported from CrispyForms itself
     "Div", "Field", "HTML", "Layout", "ButtonHolder",
@@ -67,3 +69,29 @@ class Column(Div):
     """
 
     css_class = 'column'
+
+
+class IconField(Field):
+    template = "%s/layout/input_with_icon.html"
+
+    def __init__(self, field, icon_prepend=None, icon_append=None, *args, **kwargs):
+        self.field = field
+        self.icon_prepend = icon_prepend
+        self.icon_append = icon_append
+        super().__init__(*args, **kwargs)
+
+    def render(self, form, form_style, context, template_pack=TEMPLATE_PACK,
+               extra_context=None, **kwargs):
+        extra_context = extra_context.copy() if extra_context is not None else {}
+        extra_context.update({
+            'icon_prepend': self.icon_prepend,
+            'icon_append': self.icon_append,
+        })
+        template = self.get_template_name(template_pack)
+
+        return render_field(
+            self.field, form, form_style, context,
+            template=template,
+            template_pack=template_pack, extra_context=extra_context, **kwargs
+        )
+
